requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pyhd8ed1ab_0
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h2bbff1b_1
astor=0.8.1=pyh9f0ad1d_0
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
autoencoder=0.0.8=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py37hec80d1f_5
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
category-encoders=1.3.0=pypi_0
certifi=2021.10.8=py37h03978a9_1
cffi=1.15.0=py37h2bbff1b_0
charset-normalizer=2.0.9=pypi_0
click=8.0.3=py37h03978a9_1
click-spinner=0.1.10=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cornac=1.14.1=pypi_0
cryptography=36.0.0=py37h21b164f_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
filelock=3.4.0=pypi_0
fonttools=4.28.5=py37hcc03f2d_0
freetype=2.10.4=hd328e21_0
gast=0.5.3=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.42.0=py37hc60d5dd_0
h5py=3.2.1=nompi_py37he280515_100
hdf5=1.10.6=nompi_he0bbb20_101
huggingface-hub=0.2.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.2=h0e60522_0
idna=3.3=pyhd3eb1b0_0
imbalanced-learn=0.8.1=pypi_0
importlib-metadata=4.10.0=py37h03978a9_0
importlib_metadata=1.5.0=py37_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.1=py37haa95532_1
ipython=7.30.0=pypi_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py37_7
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py37h8c56517_1
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libclang=11.1.0=default_h5c34c98_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.19.1=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.0=h2bbff1b_0
lightfm=1.16=pypi_0
lightgbm=3.3.1=pypi_0
llvmlite=0.37.0=pypi_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=py37h2bbff1b_0
matplotlib=3.4.3=pypi_0
matplotlib-base=3.5.1=py37h4a79c79_0
matplotlib-inline=0.1.3=pypi_0
memory-profiler=0.60.0=pypi_0
missingno=0.5.0=pypi_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
mpmath=1.2.1=pypi_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
nltk=3.6.5=pypi_0
notebook=6.4.6=py37haa95532_0
numba=0.54.1=pypi_0
numpy=1.20.3=py37ha4e8547_0
numpy-base=1.20.3=py37hc2deb75_0
olefile=0.46=py37_0
openssl=1.1.1l=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.4=pypi_0
pandocfilters=1.4.3=py37haa95532_1
parso=0.8.3=pypi_0
patsy=0.5.2=pypi_0
pickle5=0.0.12=py37hcc03f2d_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37hd45dc43_0
pip=21.2.4=py37haa95532_0
powerlaw=1.5=pypi_0
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.24=pypi_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.1=py37hf2a7229_1
psutil=5.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydocumentdb=2.3.5=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pymanopt=0.2.5=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.6=pyhd8ed1ab_0
pyqt=5.12.3=py37h03978a9_8
pyqt-impl=5.12.3=py37hf2a7229_8
pyqt5-sip=4.19.18=py37hf2a7229_8
pyqtchart=5.12=py37hf2a7229_8
pyqtwebengine=5.12.1=py37hf2a7229_8
pyreadline=2.1=py37h03978a9_1005
pyrsistent=0.18.0=py37h196d8e1_0
pysocks=1.7.1=py37_1
python=3.7.6=h60c2a47_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.10.1=py3.7_cuda10.2_cudnn7_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pywin32=302=pypi_0
pywinpty=0.5.7=py37_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=py37hd77b12b_2
qt=5.12.9=h5909a2a_4
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
recommenders=0.7.0=pypi_0
regex=2021.11.10=pypi_0
requests=2.26.0=pyhd3eb1b0_0
retrying=1.3.3=pypi_0
sacremoses=0.0.46=pypi_0
scikit-learn=0.24.2=pypi_0
scikit-surprise=1.1.1=py37hec80d1f_2
scipy=1.7.2=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37haa95532_0
shellingham=1.4.0=pyh44b312d_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
src=0.1.0=dev_0
statsmodels=0.13.1=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h1f41ff6_0
tensorflow-base=1.14.0=py37hc8dfbb8_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
termcolor=1.1.0=py_2
terminado=0.9.4=py37haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h2bbff1b_0
tokenizers=0.10.3=pypi_0
torchaudio=0.10.1=py37_cu102
torchmetrics=0.6.2=pyhd8ed1ab_0
torchtext=0.11.1=py37
torchvision=0.11.2=py37_cu102
tornado=6.1=py37hcc03f2d_2
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.12.5=pypi_0
typer=0.4.0=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
unicodedata2=14.0.0=py37hcc03f2d_0
urllib3=1.26.7=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.5.1=py37_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
winpty=0.4.3=4
wrapt=1.13.3=py37hcc03f2d_1
xz=5.2.5=h62dcd97_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0

pandas~=1.3.4
tabulate~=0.8.9
numpy~=1.20.3
scipy~=1.7.2
nltk~=3.6.5
typer~=0.4.0
regex~=2021.11.10
pywin32~=302
matplotlib~=3.4.3
click~=8.0.3
torchmetrics~=0.6.2
tqdm~=4.62.3
scikit-learn~=0.24.2